Organized the navigation stack, added c
