+++ src/components/GameScreen/MonsterView.js
@@ -1,0 +1,6 @@
+import react from 'react';
+
+export default class MonsterView extends React.Component {
+    // add monsters here as sub components 
+    // pass in monster power/type/etc based on character data from props
+}
