+++ src/models/Character.js
@@ -1,0 +1,8 @@
+
+export default {
+    level: 1,
+    hitpoints: 10,
+    takeDamage: (damage) => {
+        return this.hitpoints - damage;
+    },
+}
